+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 
 ## [In Development](https://github.com/ppfeufer/eve-online-killboard-widget/tree/development)
 [Full Changelog](https://github.com/ppfeufer/eve-online-killboard-widget/compare/v0.24.1...development)
-- in development
+### Changed
+- implemented a reset to defaults method in Swagger class to make ESI call handling a bit easier to code
 
 ## [0.24.1](https://github.com/ppfeufer/eve-online-killboard-widget/releases/tag/v0.24.1) - 2018-09-26
 [Full Changelog](https://github.com/ppfeufer/eve-online-killboard-widget/compare/v0.24.0...v0.24.1)